[cronopio] port patches: i64/f64-free codegen + D_DoomMain loop split
Minimal, marked [cronopio] changes to make Crispy Doom 7.1 translate to and run
on CronoVM (no i64/f64 SSA, no SDL/OS, frame-callback model):

- d_main.c: split D_DoomMain's init from the infinite D_DoomLoop (added
  D_DoomInitLoop + a non-static doom_tick the cart drives per frame); dense
  mission/banner/gameversion tables (the translator serializes pointer arrays,
  but not `ptr poison` holes); ROM IWAD path.
- m_fixed.h/.c: FIXED2DOUBLE->float; CVM_CrossDiv/CVM_ProjDiv (i64-free 64/32
  long division) used by r_segs/p_setup.
- r_main/.h, r_segs, r_things/.h, r_bsp, r_state.h, r_data, tables.c: WiggleFix
  and projection math reworked to 32-bit via cvm_qmul_16_16/CVM_CrossDiv;
  field-wise cliprange copy (avoid i64 struct copy); double color literals->int.
- g_game, p_bexptr, p_enemy, p_setup, st_stuff: angle/coord math i64-free;
  function splits to fit the register budget; float (not double) locals.
- m_config: float parse/save without atof / %f varargs.
- doomtype.h: NORETURN->empty; CRON_NOINLINE. i_input.h, crispy.h, hu_stuff,
  m_menu, v_trans, v_video: helper decls, dense switch tables, float swaps.

+++ src/doom/r_state.h
@@ -101,7 +101,7 @@ typedef struct localview_s
     angle_t oldticangle;
     angle_t ticangle;
     short ticangleturn;
-    double rawangle;
+    float rawangle; // [cronopio] f64 -> f32 (translator rejects double)
     angle_t angle;
 } localview_t;
