+++ src/i_input.h
@@ -36,8 +36,9 @@ extern int novert; // [crispy]
 extern int runcentering; // [crispy]
 
 // [crispy]
-double I_AccelerateMouse(int val);
-double I_AccelerateMouseY(int val);
+// [cronopio] f64 -> f32: CronoVM translator rejects double.
+float I_AccelerateMouse(int val);
+float I_AccelerateMouseY(int val);
 
 void I_BindStrifeInputVariables(void); // [crispy]
 void I_BindInputVariables(void);
